requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
lifelines>=0.27.8
pyarrow>=15.0.0
joblib>=1.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
# Optional; comment out if wheels not available on your base image
scikit-survival>=0.22.2
datasets>=3.1.0   # ✅ HuggingFace dataset support
python-multipart
SimpleITK
joblib
pillow
torch
torchvision
pillow
tifffile
imageio
scikit-image
scipy
openslide-python